feat: show the black hole on the empty-search state
Extract the WGSL black hole (canvas + shader + ring placeholder/fallback +
WebGPU setup) out of NotFound into a reusable BlackHole component, and render
it on the home page's no-results state — without the "4 4" digits — so an
empty search fills the dead space with the animation instead of a lone line of
text. NotFound now composes the same component as its "0".
